Function rpeg::value_conversion::rgb_to_ypbpr

pub fn rgb_to_ypbpr(
    current_img: &Vec<Rgb>,
    rgb_float_img: &Vec<RgbFloat>,
    width: u32,
    height: u32
) -> Vec<Ypbpr>

Function: rgb_to_ypbpr -> Vec<Ypbpr>

The rgb_to_ypbpr function recieves paramters: current_img: &Vec<csc411_image::Rgb> -> data type from csc411_image rgb_float_img: &Vec<RgbFloat> -> reference to the vector of Rgb float values width: u32 -> value of the current_image’s width height: u32 -> value of the current_image’s height

The purpose of this function is to calculate the Y, P_b, and P_r values and return a vector. The returning vector is a vector of Ypbpr data attributes of every pixel within the current_img passed to the function. The calculations that were provided can be found through the assignment’s documentation outline.